+++ src/providers/formattingProvider.ts
@@ -28,9 +28,7 @@ export function formatDocument(fileName: string, text: string, tabspace: integer
   /* Then format semicolons */
   doc = addNewlinesAfterSemicolons(doc);
   /* Then start by formatting patterns - beautification */
-  if (patterns.length > 0) {
-    doc = alignTextElements(doc, patterns);
-  }
+  doc = alignTextElements(doc, patterns);
 
 
   /* The make it process line by line */